Added FactorioRoller exe distribution
Included in the folder is everything for the exe to run except for a Factorio instance inside a folder called "Factorio_x64_1.1.62" with its mods folder replaced by the included mods folder here. Factorio 1.1.62 should be downloaded from Factorio.com and unzipped to a folder by that name and moved to this folder. Then cut and paste the mods folder from here to the "Factorio_1.1.62" folder, replacing the existing mods folder.